Brief test code for sending a world snapshot to clients
Wrote very brief code that parses entities into byte[] to be sent to clients to be unwrapped and instantiated, currently many problems occure, such as clients being unable to interpolate with their client-side physics with maths going to infinity for unknown reason.

Code cleanup needs to be done in the byte parsing.